Modificaciones de envio de mail, no terminado aun

+++ AppAPL.Core/AppAPL.Dto/Promocion/PromocionDTO.cs
@@ -421,9 +421,10 @@ namespace AppAPL.Dto.Promocion
 
     public class AcuerdoDTO
     {
-        public long IdAcuerdo { get; set; }
+        public int IdAcuerdo { get; set; }
         public decimal PorcentajeDescuento { get; set; }
         public decimal ValorComprometido { get; set; }
+        
     }
 
     public class SegmentoDTO
